Update jobs xcode to 16.4 (#797)
Some jobs started failing after the default machine changed to m4pro,
which doesn't have xcode 14. This updates the xcode version to try to
make them pass again

This doesn't include the iOS integration tests that are still running on
m1. I tried updating to m4 + xcode 16.4 but they are failing in what
seems to be some issue with paramedic. I tried updating but didn't help,
so I'm leaving those in m1 for now.